:recycle: Refactor: verification 패키지 구분 (#5)
